[Refactor/#135] WithdrawalScreen, WithdrawalViewModel을 screens.withdrawal 패키지로 이동
